+++ tarantool/src/ffi/sql.rs
@@ -4,7 +4,14 @@ use libc::{iovec, size_t};
 use std::cmp;
 use std::io::Read;
 use std::mem::MaybeUninit;
+use std::ops::Range;
 use std::os::raw::{c_char, c_int, c_void};
+use std::ptr::{null, NonNull};
+use tlua::LuaState;
+
+use crate::tuple::Tuple;
+
+use super::tarantool::BoxTuple;
 
 pub const IPROTO_DATA: u8 = 0x30;
 
@@ -12,6 +19,11 @@ pub const IPROTO_DATA: u8 = 0x30;
 // even if they're only used in this file. This is because the `define_dlsym_reloc`
 // macro doesn't support private function declarations because rust's macro syntax is trash.
 crate::define_dlsym_reloc! {
+    pub(crate) fn port_destroy(port: *mut Port);
+    pub(crate) fn port_c_create(port: *mut Port);
+    pub(crate) fn port_c_add_tuple(port: *mut Port, tuple: *mut BoxTuple);
+    pub(crate) fn port_c_add_mp(port: *mut Port, mp: *const c_char, mp_end: *const c_char);
+
     pub(crate) fn cord_slab_cache() -> *const SlabCache;
 
     pub(crate) fn obuf_create(obuf: *mut Obuf, slab_cache: *const SlabCache, start_cap: size_t);
@@ -131,7 +143,7 @@ impl Read for ObufWrapper {
 
 // TODO: ASan-enabled build has a different layout (obuf_asan.h).
 #[repr(C)]
-pub(crate) struct Obuf {
+pub struct Obuf {
     _slab_cache: *const c_void,
     pub pos: i32,
     pub n_iov: i32,
@@ -149,5 +161,210 @@ pub(crate) struct Obuf {
 impl Drop for Obuf {
     fn drop(&mut self) {
         unsafe { obuf_destroy(self as *mut Obuf) }
+    }
+}
+
+#[repr(C)]
+pub struct PortVTable {
+    pub dump_msgpack: unsafe extern "C" fn(port: *mut Port, out: *mut Obuf),
+    pub dump_msgpack_16: unsafe extern "C" fn(port: *mut Port, out: *mut Obuf),
+    pub dump_lua: unsafe extern "C" fn(port: *mut Port, l: *mut LuaState, is_flat: bool),
+    pub dump_plain: unsafe extern "C" fn(port: *mut Port, size: *mut u32) -> *const c_char,
+    pub get_vdbemem: unsafe extern "C" fn(port: *mut Port, size: *mut u32) -> *const c_void,
+    pub destroy: unsafe extern "C" fn(port: *mut Port),
+}
+
+#[repr(C)]
+#[derive(Debug)]
+pub struct Port {
+    pub vtab: *const PortVTable,
+    _pad: [u8; 68],
+}
+
+impl Port {
+    /// # Safety
+    /// The caller must be sure that the port was initialized with `port_c_create`.
+    pub unsafe fn mut_port_c(&mut self) -> &mut PortC {
+        unsafe { NonNull::new_unchecked(self as *mut Port as *mut PortC).as_mut() }
+    }
+}
+
+impl Port {
+    pub fn zeroed() -> Self {
+        unsafe {
+            Self {
+                vtab: null(),
+                _pad: std::mem::zeroed(),
+            }
+        }
+    }
+}
+
+impl Drop for Port {
+    fn drop(&mut self) {
+        if !self.vtab.is_null() {
+            unsafe { port_destroy(self as *mut Port) }
+        }
+    }
+}
+
+#[repr(C)]
+union U {
+    tuple: NonNull<BoxTuple>,
+    mp: *const u8,
+}
+
+#[repr(C)]
+struct PortCEntry {
+    next: *const PortCEntry,
+    data: U,
+    mp_sz: u32,
+    tuple_format: *const c_void,
+}
+
+#[repr(C)]
+pub struct PortC {
+    pub vtab: *const PortVTable,
+    first: *const PortCEntry,
+    last: *const PortCEntry,
+    first_entry: PortCEntry,
+    size: i32,
+}
+
+impl Drop for PortC {
+    fn drop(&mut self) {
+        unsafe { port_destroy(self as *mut PortC as *mut Port) }
+    }
+}
+
+impl Default for PortC {
+    fn default() -> Self {
+        unsafe {
+            let mut port = std::mem::zeroed::<PortC>();
+            port_c_create(&mut port as *mut PortC as *mut Port);
+            port
+        }
+    }
+}
+
+impl PortC {
+    pub fn add_tuple(&mut self, tuple: &mut Tuple) {
+        unsafe {
+            port_c_add_tuple(
+                self as *mut PortC as *mut Port,
+                tuple.as_ptr() as *mut BoxTuple,
+            );
+        }
+    }
+
+    /// # Safety
+    /// The caller must ensure that the `mp` slice is valid msgpack data.
+    pub unsafe fn add_mp(&mut self, mp: &[u8]) {
+        let Range { start, end } = mp.as_ptr_range();
+        unsafe {
+            port_c_add_mp(
+                self as *mut PortC as *mut Port,
+                start as *const c_char,
+                end as *const c_char,
+            );
+        }
+    }
+
+    pub fn iter(&self) -> PortCIterator {
+        PortCIterator::new(self)
+    }
+}
+
+#[allow(dead_code)]
+pub struct PortCIterator<'port> {
+    port: &'port PortC,
+    entry: *const PortCEntry,
+}
+
+impl<'port> From<&'port PortC> for PortCIterator<'port> {
+    fn from(port: &'port PortC) -> Self {
+        Self::new(port)
+    }
+}
+
+impl<'port> PortCIterator<'port> {
+    fn new(port: &'port PortC) -> Self {
+        Self {
+            port,
+            entry: port.first,
+        }
+    }
+}
+
+impl<'port> Iterator for PortCIterator<'port> {
+    type Item = &'port [u8];
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.entry.is_null() {
+            return None;
+        }
+
+        // The code was inspired by `port_c_dump_msgpack` function from `box/port.c`.
+        let entry = unsafe { &*self.entry };
+        self.entry = entry.next;
+        if entry.mp_sz == 0 {
+            let tuple_data = unsafe { entry.data.tuple.as_ref().data() };
+            return Some(tuple_data);
+        }
+        let mp_data = unsafe { std::slice::from_raw_parts(entry.data.mp, entry.mp_sz as usize) };
+        Some(mp_data)
+    }
+}
+
+#[cfg(feature = "picodata")]
+#[cfg(feature = "internal_test")]
+mod tests {
+    use super::*;
+    use crate::offset_of;
+
+    #[crate::test(tarantool = "crate")]
+    pub fn test_port_definition() {
+        let lua = crate::lua_state();
+        let [size_of_port, offset_of_vtab, offset_of_pad]: [usize; 3] = lua
+            .eval(
+                "local ffi = require('ffi')
+            return {
+                ffi.sizeof('struct port'),
+                ffi.offsetof('struct port', 'vtab'),
+                ffi.offsetof('struct port', 'pad')
+            }",
+            )
+            .unwrap();
+
+        assert_eq!(size_of_port, std::mem::size_of::<Port>());
+        assert_eq!(offset_of_vtab, offset_of!(Port, vtab));
+        assert_eq!(offset_of_pad, offset_of!(Port, _pad));
+    }
+
+    #[crate::test(tarantool = "crate")]
+    pub fn test_port_c_definition() {
+        let lua = crate::lua_state();
+        let [size_of_port_c, offset_of_vtab,
+             offset_of_first, offset_of_last,
+             offset_of_first_entry, offset_of_size]: [usize; 6] = lua
+            .eval(
+                "local ffi = require('ffi')
+            return {
+                ffi.sizeof('struct port_c'),
+                ffi.offsetof('struct port_c', 'vtab'),
+                ffi.offsetof('struct port_c', 'first'),
+                ffi.offsetof('struct port_c', 'last'),
+                ffi.offsetof('struct port_c', 'first_entry'),
+                ffi.offsetof('struct port_c', 'size')
+            }",
+            )
+            .unwrap();
+
+        assert_eq!(size_of_port_c, std::mem::size_of::<PortC>());
+        assert_eq!(offset_of_vtab, offset_of!(PortC, vtab));
+        assert_eq!(offset_of_first, offset_of!(PortC, first));
+        assert_eq!(offset_of_last, offset_of!(PortC, last));
+        assert_eq!(offset_of_first_entry, offset_of!(PortC, first_entry));
+        assert_eq!(offset_of_size, offset_of!(PortC, size));
     }
 }